Update KiBot files

+++ summer_2022_refurb/.kibot/cleanup.sh
@@ -2,19 +2,18 @@
 
 boards=("$@")
 
+cd ../
+
 for board in "${boards[@]}"
 do
     echo "Deleting documentation for $board"
     cd $board
 
-    if [ -d docs ];
-    then
-        rm -r docs
-        rm -f index.html
-        rm -f *screencast.ogv
-    else
-        echo "No docs found for $board"
-    fi
+    [ ! -d docs ] || rm -r docs
+    [ ! -d .kibot ] || rm -r .kibot
+    [ ! -e index.html ] || rm -f index.html
+    [ ! -e *-drc.txt ] || rm -f *-drc.txt
+    [ ! -e *-erc.txt ] || rm -f *-erc.txt
 
     cd ../
 done
